Web App/requirements.txt
matplotlib==3.2.2
numpy
panda
PyMySQL
Django==3.0.8
networkx==2.4
python_dateutil==2.8.1
